refactor: move useSidebar to own file in hooks
